move uopgraph to file [run_process_replay]
