+++ test/test_uops_stats.py
@@ -2,7 +2,8 @@ import unittest
 from tinygrad import Tensor
 from tinygrad.engine.schedule import create_schedule
 from tinygrad.engine.realize import lower_schedule_item
-from tinygrad.codegen.uops import UOpGraph, UOps, UOp
+from tinygrad.codegen.uops import flops_mem, UOps, UOp
+from tinygrad.codegen.uopgraph import UOpGraph
 from tinygrad.ops import BinaryOps, TernaryOps
 from tinygrad.dtype import dtypes
 
@@ -74,7 +75,7 @@ class TestUOpsStats(unittest.TestCase):
     u4 = UOp(UOps.ALU, dtypes.int, (u1,u2,u3), TernaryOps.MULACC)
     uops_fma = UOpGraph([u4])
 
-    self.assertEqual(uops.flops_mem(), uops_fma.flops_mem())
+    self.assertEqual(flops_mem(uops.uops), flops_mem(uops_fma.uops))
 
 
 if __name__ == '__main__':
